refactor policy validation
