+++ cmd/cli/kubectl-kyverno/oci/oci_push.go
@@ -14,8 +14,8 @@ import (
 	"github.com/kyverno/kyverno/cmd/cli/kubectl-kyverno/utils/common"
 	"github.com/kyverno/kyverno/pkg/config"
 	"github.com/kyverno/kyverno/pkg/openapi"
-	policyvalidation "github.com/kyverno/kyverno/pkg/policy"
 	policyutils "github.com/kyverno/kyverno/pkg/utils/policy"
+	policyvalidation "github.com/kyverno/kyverno/pkg/validation/policy"
 	"github.com/spf13/cobra"
 	"go.uber.org/multierr"
 	"sigs.k8s.io/controller-runtime/pkg/log"
